// lava-shack/admin/src/domain/system-config/types/enums.ts
export enum ConfigType {
  Payment = "payment",
  Email = "email",
  Sms = "sms",
  GoogleApi = "google-api",
  Tracking = "tracking",
}

export enum ProviderType {
  MPAY = "mpay",
  STRIPE = "stripe",
  SQUARE = "square",
  MAILGUN = "mailgun",
  TEXTGRID = "textgrid",
  GOOGLE = "google",
  MEDUSA = "medusa",
  DUMMY = "dummy",
}